Idea: Fix path normalisation
When a build file imported external modules, the `idea` command
could fail with an exception thrown by `relativize()`. This commit
addresses various corner cases and consolidates the logic for path
normalisation.

Also, `BuildConfig` was changed to prefer a relative build project
path as this value gets propagated to the generators.