BuildConfig: Detect target platforms of test module if unspecified
Currently, the logic in `BuildConfig` only considers the
platform-specific modules defined on a test module. Change the logic
to inherit all targets from the parent module.

Change the type of `import` on `Build` to `Path` to avoid the
duplicated normalisation logic of `fixPath()`.

Also, break `load()` on `Build` apart into separate functions to
allow for better testing.