Implement custom build targets
Up until now, Seed users had to write shell scripts in order to
build non-Scala artefacts. This proved difficult when artefacts had
to be built in multiple dependent modules and copied over to the
root project's build path.

These changes allow to define custom build targets on any module. A
custom build target can run a command or class. The execution is
synchronous or asynchronous. Environment variables are defined which
allow the command/class to access the build path and module's project
path. For each build target, an IDEA module is generated if the root
path is set. Build targets can run be run in watch mode.

Further changes:
- Introduce a new CLI command `build`, similar to `link`. Both
  execute build targets transitively.
- Fix `seed link` command when `--server` is omitted. The process
  terminated immediately without waiting for Bloop to finish.

For more information, please refer to the updated README.