Bloop: Fix output path of submodules
Currently, the output path of a submodule includes the relative path
to its build file, although it should only be relative to the root
project's build path.

Another limitation that these changes address is that the output
folder was not created by Seed such that Bloop might fail to link
the project.