Use ZIO instead of Futures
The motivation for using ZIO is that it propagates errors better.
For example, if any of the executed commands fails, the entire
execution pipeline should be interrupted.

Previously, we relied on a simple workaround to implement this
behaviour by calling `sys.exit()`. However, when Seed is running in
server mode, any failing command will shut down the server which is
undesired.