Inherit settings when loading build file
Centralise the logic for inheriting settings from the project
section and base modules. The inheritance step is now performed when
loading the build file. As a consequence, many functions do not need
to take a reference to the project definition or parent modules
anymore. This greatly simplifies the code for generating Bloop/IDEA
configurations, and makes the logic less error-prone.

The build file's project section is now entirely optional and its
values can be set on modules instead. Now, Scala options can be
extended in platform modules, for example to set additional flags
required by the Scala.js compiler such as
`-P:scalajs:sjsDefinedByDefault`.

Furthermore, every module is compiled with the Scala version and the
options it was defined with. Therefore, it is not possible to depend
on modules anymore which have an incompatible Scala version. For
example, including a 2.12 module in a 2.13 module will trigger an
error.

Closes #18.