BuildTarget: Expose source paths as environment variable
It is possible for multiple modules to share the same build target, but
the executed class/command could not determine which modules referenced
it. This limitation is addressed by introducing a new environment
variable `MODULE_SOURCE_PATHS` which will contain the source paths of
all modules that depend on the build target.